https://github.com/appleboy/gorush/issues/668
GoRush uses the KeyPath file extension in order to determine how to create the certificate (.p12, .pem, .p8). However if you use Terraform to deploy GoRush, the name of volume_mounts can only contain alphanumeric characters, hyphens and underscores but no dots. This seems to make it impossible to deploy GoRush with Terraform and mount a KeyPath file from the GCP Secret Manager.

This change implements the suggestion written in the #668 ticket: if KeyPath has no extension, we can fallback to KeyType to determine the type.